+++ assets/scripts/app.ts
@@ -1,4 +1,5 @@
 import '../styles/app.css';
+import '@awesome.me/fa-pro/all.js';
 import ComponentLoader from './Core/ComponentLoader';
 import FlashHandler from './Core/FlashHandler';